rewrite fastify with new plugin system
