+++ packages/datadog-instrumentations/index.js
@@ -10,6 +10,7 @@ require('./src/cucumber')
 require('./src/dns')
 require('./src/elasticsearch')
 require('./src/express')
+require('./src/fastify')
 require('./src/find-my-way')
 require('./src/generic-pool')
 require('./src/google-cloud-pubsub')
