+++ packages/datadog-plugin-fastify/src/index.js
@@ -1,6 +1,19 @@
 'use strict'
 
-module.exports = [].concat(
-  require('./fastify'),
-  require('./find-my-way') // TODO make this its own plugin, since restify uses it too
-)
+const RouterPlugin = require('../../datadog-plugin-router/src')
+
+class FastifyPlugin extends RouterPlugin {
+  static get name () {
+    return 'fastify'
+  }
+
+  constructor (...args) {
+    super(...args)
+
+    this.addSub('apm:fastify:request:handle', ({ req }) => {
+      this.setFramework(req, 'fastify', this.config)
+    })
+  }
+}
+
+module.exports = FastifyPlugin
